`ora2pg`: a script for converting Oracle DDL to PostgreSQL DDL
The idea is that `ora2pg` will run on the `.ddl.txt` files generated by
`build-schemas.ps1` to generate a schema that can be run against the
local PostgreSQL instance.

That same schema can then be used to help initialize the remote database
on EC2 (minus the foreign table clauses, of course).