Add all tables under 1M rows in target set to FLOW.config.json
This covers everything except CNT_DET (126M rows) and CNT_SPD (13M
rows), plus ACC (4.5M rows) on the CRASH side.

To properly separate the FLOW and CRASH replication jobs, we'll need to
make sure the relevant files don't collide.  One way is to use the
`$config` parameter as a prefix to `".config.json"` for the config
filename, then to use it as a suffix on `"flashcrow"` for the working
folder and archive names.

We're already passing this parameter to `replicator-transfer.sh` via our
`plink` invocation, so we can simply add the necessary suffix to paths
there.