Integrating `shellcheck`, a static linting tool for Bash scripts
so that we can get proper static linting for all target languages.

Also added this to `scoop-requirements.ps1` (and tested that change by
running it!), and updated `package.json` with the appropriate
`lint-staged` declaration.  As an added bonus, I've updated VSCode
workspace settings to integrate this with the `shellcheck` extension
there.