Add unit testing to pre-commit hook
This resulted in an epic journey into dependency hell and back, in which
I eventually found out that `@vue/cli-plugin-unit-jest` plays poorly
with Babel 7.  As a workaround, I instead call `jest` directly, and
set the necessary `vue-jest` environment variables in `jest.config.js`.

This took *way* longer than it needed to (~2 hours), but now we can eat
our cake and run tests too.