Fix bugs in replication script
Use `Set-Location` to do everything relative to `scripts/replication`.
Add `-s` flag to `curl.exe` to squelch output on stderr, which seems to
confuse `$ErrorActionPreference` when run as a scheduled job.  Add `-q`
flag to `dos2unix`, and squelch `NOTICE` messages from `psql`.

Also allow user to supply either an IP address or a stack ID when
calling `replicator-local.ps1`, so that we can also support replication
to BDITTO's machine.