Fix REST API tests
so that these can run under our current OpenID Connect authentication
setup.

The "fix" is actually more of a workaround: instead of going through the
OpenID Connect flow before creating an application session, we add a new
endpoint that can be used in testing / development environments to
authenticate a test user.

I investigated crafting a valid JWT and moving this to a Cypress-based
end-to-end test, but eventually settled on this workaround approach -
after all, we're not trying to test OpenID Connect here!