Vagrant configuration and provisioning scripts
So far, this allows us to install the application, run it, and access it from the host machine.  It also allows us to run the pre-commit hook.

It doesn't yet set up the database, nor does it allow Visual Studio Code to tunnel into the machine to edit files there.  Once we do both of these, we'll have a standalone development environment that can be deployed to any host operating system.