Layout fixes
Replaced `flex-grow` with `flex` so that display / map panes share space better with each other.  Removed a couple of layers of `<div>` nesting in the layout, and fixed an issue where selecting enough counts would overflow the browser window - now the footer is fixed as expected, and the display pane contents remains fixed as well.  Vertically-aligned checkboxes for day of week, so that these appear in a way that more closely mirrors the calendar.