Custom checkbox and toggle
using CSS - I initially attempted to build these using Vue components, but quickly found that this was going to be more work than expected.  Scaled back that work to just using the CSS, and leaning on HTML templating + props to do the job.  (It's a bit more verbose than having an all-in-one `TdsCheckbox` component, but it's probably also more readable from a "this just HTML" perspective.)