Implement date range filtering
The hardest part here is handling "no existing count" categories when filters are active: in some cases (e.g. the count type is shown), we want to show the non-existence of a count, whereas in others (e.g. date range filtering) we don't.

While we could have stuck with the previous "dummy count" approach, we would have had to move it outside `SampleData` soon enough - that's what this change effectively does.