Move new form elements into FcDetails components
as before.  Also fixed a couple of minor spacing, width, and copy issues.  As part of this, I've moved study request and per-study metadata back into the Vuex store, and used state / mutations to access / update that from the `FcDetails` components.

Next up: pass this state into a redesign of the Confirm screen.