Add indexes capability to replication pipeline
In some cases, we want to add indexes that weren't in the original
Oracle tables!  To do that, we add a per-table `"indexes"` key to the
JSON config files.

This key is an array of arrays of column names.  For each array of
column names, we generate an index on those columns of the current
table, using an auto-generated name based on the table and column names.

Similar to `"exclude"`, we use `replication_utils.get_table_config()`
to provide a default empty array via `setdefault()`.

I tested this quickly by inserting an `Exit-Success` statement after
JSON / SQL generation is finished, running, and inspecting the output.
I'll also be monitoring replication over the next 24 hours to ensure
that the tables have indexes as expected.