Chunked loading of data into PostgreSQL RDS
and approximate count checking, as opposed to the old `COUNT(*)`
approach - this should save us considerable time in replication.

It seems that the size of `CNT_DET` is a bit much to load into our RDS
in a single transaction!

Running a test on the remote side of FLOW replication right now, will
resume both CRASH and FLOW jobs before EOD and monitor between now and
launch.