Reorganize Jest tests
so that only unit tests are run during pre-commit, and so that we can run different sets of test (unit, REST API, database) using the `TEST_DIR` environment variable.

For convenience, I've added a *bunch* of npm scripts to run the backend, frontend, CI tasks, and various granularities of testing.