Formalize target browsers list and use for JS / CSS compilation
As part of this, I've used `postcss-custom-properties` with our CSS custom properties ("variables") broken out into a new file `@/components/tds/tds.css`.  This allows us to support browsers that don't fully support CSS custom properties.

I've also fixed a small but important UX issue in Firefox: when logging in, it would redirect the user to `/api/auth/stub`, but never go back to the homepage.  I tracked this down to a missing redirect on that endpoint in `server.js`.