Fix padding issue in navbar on Firefox, Edge
It turns out that `margin-block-end` et al are experimental, and that the correct cross-browser way to do this is `padding-left: 0`.  Much simpler!

With this fix, we can also drop `postcss-custom-properties` from our `package.json` and PostCSS config - after dropping IE 11 support, we don't have to deal with transforming these.  (Just as well for now, because adding this transformation doubles our build / hot-reload time!)