FcToast: a lightweight toast component
so that we can show non-blocking messages to the user just below the topbar.  This is less intrusive than a full modal.

`FcToast` allows success, warning, and error variants.  The success variant should be used to notify the user that an operation has succeeded.  The error variant should be used if something failed that would substantially impact use (e.g. we failed to write to database, a system component is down, or other operational failures).  The warning variant is used for in-between "soft feedback" (e.g. the user cannot proceed until they do something further, a feature is still being worked on, etc.) - in other words, there may be a failure, but the user impact is expected to be minimal.

Under no circumstances should an error message indicate success.  For example: say the user cancels an operation; you should use success to indicate that it was successfully cancelled.

`FcToast` messages should be short, readable, and informational.

`FcToast` should *not* be used if the user needs to make a decision, or if they need to perform follow-up actions / interactions.  Simple yes / no (or OK / cancel) decisions are a good use case for `TdsConfirmDialog`; anything more complicated should use a component built around `TdsModal`.  Keep in mind that the toast disappears after a few seconds, and that the user may not even read it - they're more likely to scan for icons / colour to get a quick sense of whether things are OK.  (So: don't embed links, buttons, etc. in toasts!)