TdsPanel: a component for simple informational panels
`TdsPanel` has three variants: info, warning, and error.  These variants, in order, draw progressively more attention to specific parts of the interface.

`TdsPanel` is used where you want to inform the user of something within the flow of an interface, such as the validation state of a form field.  If you just want to notify the user whether or not something succeeded, consider `FcToast`.  If you want the user to decide something, consider `TdsConfirmDialog`.

The error variant is most commonly used for form validation.  Warnings are often used to indicate that a feature is coming soon, or that the user has chosen an uncommon option requiring additional input / follow-up work (e.g. urgent priority for requests, Other hours for studies).  Info is used to provide additional information or details (e.g. count hours for Routine / School counts.)

We may add a success variant, but for now `FcToast` is the best way to indicate success of an operation.