Use database harness for UserDAO test
by calling `beforeAll()`, `afterAll()` and using `child_process.execFile()` to run our startup / shutdown scripts.

I tested that this is working by commenting out the `afterAll()` implementation, running the test, and checking at various points that the expected data was written to the database in our testing harness.  I also uncommented `afterAll()` and tested that the ramdisk is properly cleaned up at end-of-test.

My next step will be to refactor this so that we can reuse the database harness across several tests; other than that, we're ready to expand our database testing coverage!