Convert WebInitController to ES6
as a test here.  This enables us to see both what has to happen within a module, and what has to happen in all modules that `require()` that module.  It also suggests that we should start "top-down", i.e. convert entry points like `server.js`, then convert modules they `require()`, and so on.  That will minimize the amount of transitional `require().default` cruft needed.