Remove not-allowed cursor on disabled elements
There are a couple of usability issues with using this cursor:

- it signals to the user that the action is not possible, but doesn't explain whether that's true temporarily or permanently;
- the user has to hover over the element in question to see the cursor, whereas simply changing the visual appearance of the element can signal that it's disabled.

`not-allowed` was left on disabled dates on `DatePicker`, since there it has a clearer semantic meaning: it is not possible to pick the date in question.  This is also a place where the user must interact with the element to see full details in the first place, so it sends a much clearer signal there.