Set feature code everywhere we set a location
We don't have a great way yet to set the feature code for a count, but we may not have to solve that problem yet - `FcDisplayViewDataAtLocation` fetches the centreline data from our API on load, which includes the feature code.

This might become an issue if / when we enable filters in "View Map".