Add components, backend server file to docs:js target
so that we can cover *everything*, as per the issue criteria.

We can't yet include components and top-level files (e.g. `router.js`) in `pre-commit:lint-docs`, as `documentation lint` doesn't support the `--babel` option.  I spent a bit of time looking into workarounds, but am going to punt on that for now.