Add CSV / JSON reports for speed percentile
Found a bug in `ArrayStats.histogramPercentile()` in the process: when `Math.round(p * totalCount) === 0`, the `if` condition would trigger on the first bucket; if `count === 0` at that point, we'd then get a divide-by-zero error.

Updated the logic here to use floating-point `cutoffCount`, so that we can always return a reasonable value.  The special cases above that ensure `p > 0 && totalCount > 0`, which means `cutoffCount > 0`.