First steps towards Intersection Warrant Summary report
This turns out to be a harder report to implement than expected, as it relies on an understanding of segments as intersection legs and major vs. minor roads.

Nevertheless, I've got a first stab at an implementation in MOVE Reporter.  I'd like to eventually move this logic to our Airflow pipelines, but that work will most likely come in a future slice.