Test Intersection Warrant Summary report against legacy values
so that we can double-check that the logic does the right thing on actual data.

Rather than integrate this with the database layer, we dump JSON input (from our `/count/data` endpoint in the MOVE web backend) and create JSON output (from looking at legacy reports), and test the logic in isolation from there.

We can extend this same strategy to the other report types, to ensure that we have at least some coverage here.