Test TMC reports against legacy examples
which is good, as it uncovered a few bugs in our implementation thus far!  Most importantly, we tracked down (and added a test case for) a bug in `ArrayUtils.selectIndices()`, and we moved the "direction swapping" logic into `ReportCountSummaryTurningMovement` so that both TMC reports could benefit from it.

Updated the TMC summary web report as well to match - this simplifies our `v-for` loops in several places here considerably.