PDF generation of charts using d3 / SVG
This turns out to be more flexible than `pdfmake`'s `canvas` block, with a slight caveat: `pdfmake` doesn't handle CSS in SVG, so far as I can tell.  To work around that, `BarChartGenerator` takes a couple of colors (bars and axes) as options.  In the backend, we provide these colors from `MovePdfGenerator#generateBarChart`.  In the frontend, we rely on CSS.

(This also gives us the flexibility to define separate print and web styles, if need be.)