Small changes to table layouts: auto-width, tableStyle
Turns out Vue doesn't like to use `style` as a component prop name (understandable, since that conflicts with the HTML attribute), so I renamed this to `tableStyle`.

Also added `autoWidthTable`, which does nothing in `MovePdfGenerator` but instructs `FcReportTable` to add the `.auto-width` class to the table for auto-width tables.  This allows us to match the appearance of those "boolean flag" tables at the bottom of the warrant report.

I have concerns about font size in the more information-dense TMC reports - we may have to figure out something there to make these more readable.