Update vector tile generation tasks
Commented out bigdata commands for now, since we already have a copy of the centreline.  Also switched from hardcoded hostname to the `env $(xargs < ~/cot-env.config)` trick.

In the process, simplified the job a bit - instead of decompressing the tiles, we just add `Content-Encoding: gzip` to headers returned, and rely on the browser to decompress client-side.