Fix DateTime handling in DatePicker hierarchy
Tracked the study requesting issue down to problems with how we handle timestamps here - we'd changed `FcFilterDate` over to use `DateTime` instances instead of JS native `Date`, but forgot to change this in `FcDetailsStudyRequest` as well.  To prevent future issues, I've pushed this handling down to `DatePicker`, so that we can isolate all mapping between `Date` and `DateTime` in the frontend to `DatePicker` and get this behaviour automatically if we add future datepicker controls.