Refactor ArcGIS fetching script
courtesy of @chmnata, who created the initial version of the script this one is based off of.  The primary difference here is that `get_layer` now outputs SQL that can be run to populate the table, as opposed to populating the table itself.  This is intended for use in Airflow scripts, where it is important to avoid partial copies of tables from overwriting known good copies (such as can happen if, e.g., a network outage causes your data loading script to fail halfway.)