Array input for CC email addresses
Now that we're using the same interface for create and edit operations, the small ways in which the state in this interface (as originally developed) differs from the state in database (and the state sent via `studyRequestModel`) are becoming apparent.

The first and most obvious one was `ccEmails`, which in the frontend was a comma-separated list of email addresses but in the backend is an array.  This commit addresses that disparity.

The main remaining disparity is around `hasServiceRequestId`, which exists in the frontend interface but not in the backend schema.