Show chips for filters
The hard part here isn't showing the chips - it's linking them with the filters in both `FcDisplayViewDataAtLocation` and `FcDialogStudyFilters`.  To further complicate matters: `<v-dialog>` sticks around even when closed (e.g. via `v-show`), so we can't rely on prop -> data initialization.

Fortunately, there's an easy solution: don't use the activator slot in `<v-dialog>`, and use `v-if` to remove the `<v-dialog>` element completely when it's closed.