Drop Edge support for now
Need to determine strategy here.  The newer Chromium-based Edge has better ES2016+ feature support, but is incompatible with Symantec Endpoint Protection 14.0 as installed on City machines, and requires an additional install - however, this installation will be automatic in upcoming Windows Updates, so supporting the old Edge doesn't make a lot of sense longer-term.