Async loading for popup description
Right now this just uses `setTimeout(..., 500)` to fake async loading, but the idea is to use `getFeatureDesciption()` to fetch additional details from the backend.  This will allow us to show richer details without needing to constantly rebuild tiles and/or package all relevant details into the feature properties.  (In turn, this means we can keep feature layers small, which will have a beneficial effect on performance.)