Dynamically loaded popup content in PaneMapPopup
This turned out to be a larger task than expected: some of the details we needed as per designs aren't present in the feature layer, meaning we need to fetch them from database.  This in turn means loading states, `async` functions, etc.

After a bit of refactoring, though, we have a workable solution that can be extended as we add more layers :)