Persist selected popup alongside hover popups
Ran into two quirks of Vue's component model here: component reuse and root element handling.

For performance reasons, Vue tries where possible to reuse instances of the same component, and to simply rebind data on them.  However, to integrate `FcPaneMapPopup` with `mapbox-gl`'s Popup UI element, we use the `created()` and `beforeDestroy()` lifecycle hooks.  A component reused by Vue is not destroyed, so these lifecycle hooks are not triggered.  This can be fixed by adding the `:key` attribute, and by providing values such that hovered and selected popups can never have the same key; we ensure that by giving a separate prefixed namespace to both sets of popups.

It also seems (though I haven't had a chance to deep-dive on this) that Vue doesn't like root elements of components being yanked out of their position in the DOM tree.  We borrow an idea from `vue-mapbox` here: the `<v-card>` containing popup content is wrapped in a `<div class="d-none">` and given a `ref="content"` attribute.  We can then pull `this.$refs.content.$el` out of place instead, which bypasses this issue.

With these two quirks addressed, the two popup types show up as expected.  We also use `z-index` and a special class on hovered feature popups to ensure that selected popups appear on top of hovered popups, and that map controls appear on top of both.