View Data opens drawer, de-duplicate popups
One issue that arose during our recent usability test was that, once closed, the side drawer is hard to find if you don't already know it's there.  We address this here by having the "View Data" button also open the side drawer.

By showing both hovered and selected popups, we also made it possible for both popups to show on the same feature.  With a bit of reworking, we can use the feature keys previously used to help with component lifecycle issues to de-duplicate these, so that the hovered popup does not show on a feature that is already selected.