More affordances for drawer toggle, MOVE "home" link
The drawer toggle now has a large icon, better on-hover shading, and a tooltip that provides contextual information.  It's also now an `<FcButton>`, meaning that it's part of the tab order.

To aid in navigation, the MOVE icon at top-left is now a secondary way to get back to "home", i.e. the initial map view.  (We can always remap this later, e.g. to an "About" page or some tutorial / onboarding materials.)