Add hospitals to GIS layers copy and vector tiles ETL jobs
This only generates level 2 tiles - for the level 1 tiles, we need to update `dump_dev_data` to include hospitals.

Looking further at our latest dumps, though, we're actually exporting a number of materialized views - which means we get the view definition, not the underlying data!  We'll need to update the dev data dump ETL job to ensure we're getting proper datasets here.