Map layers styling: collisions circles, heatmap
with proper crossfade.  It turns out that the total radius is `circle-radius + circle-stroke-width`, so we have to adjust both based on KSI vs. non-KSI status.  Also had to be extra-careful with transitions over the heatmap layer as we move from zoom levels based on source layers to those based on `MapZoom`.