Spritemap for map layer icons
As part of this, we rip out the ESRI spritemap and replace it with our own, which means we also need to drop any layers from the ESRI styles that depend on this spritemap.  From a quick check over the style definitions, and from inspecting the rendered output at multiple zoom levels, this should be fine.