Switch dependency tools to allow for easy updates
`check-outdated` worked OK for checking which packages are outdated (with the ability to ignore packages like `openid-client`), but it didn't allow us to easily update just those packages. `npm-check-updates`, on the other hand, makes it easy to do that by running `npx ncu -u && npm install`, so I've switched over to that.