Back out "ease on drawer open" behaviour
This seemed like a good idea to deal with an issue where opening the drawer caused a race between map resizing and easing to the pin, but it has the unintended consequence of making it harder to open / close the drawer and keep your context - since that sort of thing is pretty central to how people are using this, I've backed it out.  Users can always use the new "recenter" button to achieve the same effect manually.