Confirm dialog when leaving View Reports
unless the user leaves via the "View Data" button, which takes them back to the View Data drawer at the same location.

There are a *lot* of ways to leave this page - so rather than hunt them down one by one, we use a navigation guard.  This introduces its own issues, as we need to exempt the "View Data" button from this guard.  In addition, we had to update the Vue router and `FcPaneMapPopup` to correctly handle the drawer on transitions, and to correctly show popups in View Reports.

Finally, this commit introduces general-purpose confirm and alert boxes.  We haven't used the alert box yet, but the confirm one appears to work as needed.