Fulfillment status bar for View Request
This doesn't yet support all the different request status codes, but it's a start.  The bar is built out of a `<v-progress-linear>` with several `<div>` circles, using border colours (as per designs) to create visual separation / connection as appropriate.

Also made a couple of small initial changes to info hierarchy underneath the bar, and updated the spacing on View Request to more closely match designs.