Explanatory messages around days of week
This was a frequent point of confusion with our Traffic Ops users: if I select 3 days for a 1-day count, is it performing the count on all 3 days or just 1 day as requested?  To reassure users here, we've added messages here to explain how Study Days and Study Hours / Study Duration interact.