Add search bar to Track Requests at top-right
The biggest hurdle here was styling: I haven't found anything in Vuetify like Bootstrap's input groups, which is essentially what the designs call for.  It's mostly there now, with the exception of a couple of small visual nits:

1. there's a double border where the two parts of this input control meet;
2. when clicking one part, the border on the other part expands.

2) could _arguably_ be a feature, since in some sense the whole thing is intended to be one input control.  1), however, just looks a bit weird.  I couldn't seem to track down where this border is added - all attempts to override `border-right` (or `border`) just added an additional border, which makes me think this might be some `outline` or even `box-shadow` hackery.  I also tried using `margin-right: -1px` to overlap these.  That works, but since both parts have a semi-opaque border, that leads to a darker 1px border which is still kinda weird.

Anyways: these don't affect the *functionality* of the input control, so I'm going to leave it for now, and we can circle back around to these if there's time (and anyone notices).