Item handling in Track Requests data table
We were previously just piling the location and requester info into the same object, but this led to validation issues in the backend when sending updates to `PUT /requests/study/{id}`.

Fixed that by preserving the original study request object as part of the data table item.  We still have to make sure that these have a top-level `id` field, though!