Implement filtering of items
Pretty straightforward - we build a function that checks whether a data table item matches the active filters, and use that in a computed prop to filter the set of items.

For now, the filtering implementation is entirely frontend.  As the number of study requests grows, and once we gather more user feedback on this, we'll move this to the backend similar to count summaries in View Data.