Stub out drawer and generators for collision reports
This causes "View Reports" in the View Data collisions section to take the user to a page showing the report drawer for collision reports.

To get this passing pre-commit checks, we've also added stub generators for these new report types.  They don't do anything, and in fact would probably crash if called, but we'll be working on that in upcoming pull requests.