Month-of-year breakdown
as our first example of an event-level cross-tabulation!  This also forced us to introduce some configurability in the cross-tabulation process, so that we could return the entire list of months (12, as opposed to the usual limit of 10) in calendar year order (as opposed to sorted in descending order of total).