Allow user to clear location search bar
Initially we considered a confirmation flow, but it turns out that this can be handled in all 5 screens without needing to confirm: View Map is trivial, View Data just goes back to View Map, Request Study and Edit Request flag that the user must select a location to save.  The only kind of complicated one is View Request - we don't have a great solution there yet.