Switch to multi-location state in vuex store
This is a pretty sweeping change: we have to update a whole slew of callsites, and in some cases (e.g. Request Study) we need to provide a computed property `location` so that we can revisit those later.

Nonetheless, it's a big step towards getting these features in.  Aside from corridor routing and (once we get to it) backend storage / representation of bulk study requests, most of the remaining work here is in the frontend.  (Not to say this won't be complicated as well, but that at least we can now focus changes largely within one component of MOVE.)