Show currently selected location on map as well
both to help during multi-location selection editing, and to prepare for later View Data interaction work.

I haven't implemented click handlers on the map pins themselves - I'd like to get further on this and put it in front of the team before digging in on that, to make sure it's an interaction that makes sense throughout multi-location.