Job monitoring toast
as per designs.  Most of the complexity here comes from tracking job progress - we have to create a reasonable time estimate, and it's ideal if it doesn't appear to "bounce" too much!  Since we can't know how far along an individual report request is, we instead decrement the estimated time remaining on each tick, and only update the estimate when the number of reports completed changes.