Refactor job polling logic into separate module
in preparation for reusing this logic in the download manager.  Also shored up the logic around stopping the poller - we want to avoid unnecessary requests, especially when we have several of these operating on a page.

(This points out that we may also want to rate-limit this in some way...)