Add export mode to data
and pass metadata through to the `JobMetadata` instance.  Also had to do a bit of wrangling here around the `Joi.compositeId()` validator - it turns out to be inconvenient in some cases that this validator changes the `.toString()` value, such as when we need to turn around and store that value elsewhere!