New intersection, midblock, and deselected icons in map
which are now integrated with `FcIconLocationMulti` and `GeoStyle`.  However, we currently don't set the `deselected` flag in `FcPaneMap` - we'll need a way to pass that information from Bulk Request Study to the map, probably via `vuex` state.  (The difficulty there is managing that state when the user enters / leaves the Request Study flow; outside this flow, we only want to use `locationIndex` to *select* locations, as opposed to using some `locationIndicesDeselected` state to *deselect* them.)