Show deselected locations in map
by passing these into the `vuex` store.  `FcPaneMap` is _definitely_ due for a refactor, especially considering that we plan to reuse it in View Request - but we'll have to see how much effort there makes sense right now.  This work also suggests that we should be storing the location selection _and_ deselected indices in the bulk study request object, as otherwise it will be difficult to show this in View Request.