Refactor: centrelineKey now takes a feature
where a "feature" is any object with keys `centrelineType` and `centrelineId` that between them identify a centreline feature.  This means that most of the frequently-used data types in MOVE - locations, study requests, collisions, etc. - can now be passed to this function directly without first destructuring them.  It also means that it is no longer possible to mix up the argument order, which is what prompted this refactor in the first place!