Bulk request status bar
The states here are a good deal more complicated than in the non-bulk request case, owing mostly to the fact that now the overall bulk request status might differ from the status of individual study requests.  For now, we make the tradeoff that _all_ circles where there is at least one request are shown; additionally, we will force the `ASSIGNED` circle to be shown if the bulk request has reached or passed the `ASSIGNED` stage, even if no requests are still in `ASSIGNED`.