Tweak map controls layout for better visibility on 1280x720
Moved the scale and map data attribution to bottom-left, which allows us to bump the rest of the controls down and de-clutters that bottom-right area.  Added `max-height` on the study date dropdown menu, so that it doesn't overlap the map even in vertical split-screen mode.  Finally, added a media query in `FcLayoutDrawerMap` that applies `max-height: 192px` and `overflow: auto` to the legend on screens with a height of 900px or less.

With these changes, we should at the very least minimize conflict between these controls even on 1280x720 (which is the smallest resolution we're seeing in our analytics so far).