Fix text entry bug in date picker
It turns out that `19` is a valid ISO datetime string, which is interpreted as "today at 19:00".  This meant that, whenever the user entered the first two digits of a year, it would detect that as a valid date and fill in the date picker with today's date.

Checking this against a full date regex fixes the issue.