FcTextarea wrapper component
in place of `<v-textarea>`, to capture common attributes.  Upon closer inspection, the `<fieldset>` element here is coming from `outlined`, which uses it in a presentational manner.  (The same is true of other Vuetify input controls.)

However, it also uses `aria-hidden="true"` to hide these, so perhaps an argument can be made there?