Add "other study type" field to study requests
complete with validators (both Vuelidate and Joi), a new column in the database, updates to `FcStudyRequestStudyType` to conditionally include this field, and updates to any unit tests that were broken by these changes.

Next up: we'll need to ensure `studyTypeOther` persists to database properly, is fetched from database properly, and is shown in the frontend when viewing a study request.