Show other study type details
in a few places.  The slight complication, of course, is that both studies and study requests have study types, but only the latter will have `studyTypeOther`.  To help make sense of this, I've also added a `dataAvailable` field to study types, to distinguish those that can be expected to show up in studies from those that can be requested (but not viewed).