Fix bug in TMC-related directional reports for midblocks
In the case of TMCs along midblocks, `CentrelineDAO.featuresIncidentTo` returns the incident intersections at the ends of the midblock.  This results in errors when using those incident features to infer road directions, so we force it to return the empty array in this case.  (This in turn causes `ReportBaseFlowDirectional` to fall back on North-South / East-West for roads.)