Get Peak Hour Factor side-by-side test passing
The fix turns out to be surprisingly simple: PHFs are calculated using _vehicle_ totals, not multi-modal totals.  (It's less clear whether this *ought* to be the case - aren't we trying to move to more multi-modal analyses, and treat cyclists / pedestrians as equal right-of-way users? - but for now, getting these to match with legacy systems is our priority.)