Use ReportTimeUtils for TMC summary report
so that this can finally have the same peak-hour logic as our PHF reports.  Also fixed a "bug" where adding rounded numbers doesn't give the same value as rounding added numbers.  (This is expected, but it's also surprising for users who might not know that this is what's happening under the covers.)