Overhaul of collision, common, study filters
to get rid of the differences between our `lib/model` validation schemas, our `vuex` state, and how these are used in various callsites (e.g. `GeoStyle`, `CollisionDAO`, etc.)

`NormalizeUtils` has definitely been *useful* as a stopgap measure where we need to work around quirks in the data, but over a couple of years on MOVE development it's consistently been a Good Idea to aim to get rid of things there where possible.  (These usually point to code smells where we're doing the same thing in slightly different ways in a bunch of places.)